+++ Inkybot.Dofus/Contracts/DofusSinkProvider.cs
@@ -2,6 +2,6 @@ namespace Inkybot.Dofus.Contracts
 {
     public interface DofusSinkProvider
     {
-        public float Sink { get; }
+        public decimal Sink { get; }
     }
 }
